reverting pytorch version changes

+++ ssd/layers/modules/multibox_loss.py
@@ -117,7 +117,7 @@ class MultiBoxLoss(nn.Module):
 
 		# Sum of losses: L(x,c,l,g) = (Lconf(x, c) + αLloc(x,l,g)) / N
 
-		N = num_pos.data.sum().float()
+		N = num_pos.data.sum()
 		loss_l /= N
 		loss_c /= N
 		return loss_l, loss_c
